requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.7.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
Pillow>=9.5.0
scikit-image>=0.21.0
tqdm>=4.65.0
pyyaml>=6.0
pandas>=2.0.0
jupyter>=1.0.0pytest>=7.0.0
lpips>=0.1.4
